final update from deploy

+++ repositories/users.js
@@ -2,15 +2,11 @@ const User = require('../models/user')
 
 const repositories = {
 
-    findById: async (id) => User.findById(id).exec(),
-
-    findByEmail: async (ctx) => User.findOne({ email: ctx.request.body.email }).exec(),
-
-    findUsers: async () => User.find({}).exec(),
-
-    findByIdAndDelete: async (id) => User.findByIdAndDelete(id).exec(),
-
     toClient: (user) => user.toClient(),
+
+    findById: async (id) => User.findById(id).exec(),
+    findByEmail: async (ctx) => User.findOne({ email: ctx.request.body.email }).exec(),
+    findUsers: async () => User.find().exec(),
 
     create: async (ctx) => {
         user = new User({
@@ -30,16 +26,19 @@ const repositories = {
         return user.toClient()
     },
 
-    // REFATORAR O SERVIÇO DE LANÇAMENTOS
-    /*
-    const delete = async (ctx) => {
-        const n = await Record.countDocuments({ owner: ctx.user._id }).exec()
-        if (n > 0) return ctx.status = 409
-        await User.findByIdAndDelete(ctx.user._id).exec()
-    }
-    */
+    delete: async (ctx) => {
+        const documents = await Record.countDocuments({ owner: ctx.user._id }).exec()
+        if (documents === 0) await User.findByIdAndDelete(ctx.user._id).exec()
+        return documents
+    },
 
-    list: async (ctx) => {
+    clear: async () => {
+        const documents = await Record.countDocuments().exec()
+        if (documents === 0) await User.deleteMany().exec()
+        return documents
+    },
+
+    list: async () => {
         const users = await repositories.findUsers()
         for (let i = 0; i < users.length; i++) {
             users[i] = users[i].toClient()
@@ -47,16 +46,9 @@ const repositories = {
         return users
     },
 
-    /*
-    const n = await Record.countDocuments().exec()
-            if (n > 0) return ctx.status = 409
-            await User.deleteMany().exec()
-            ctx.status = 204
-    */
-
     updateBalance: async (ctx) => {
         const user = ctx.user
-        user.balance += ctx.request.body.value
+        user.balance += parseInt(ctx.request.body.value)
         await user.save()
     }
 }
